added react router and login page
