+++ package.json
@@ -25,6 +25,7 @@
     "babel-loader": "^8.0.6",
     "css-loader": "^1.0.1",
     "style-loader": "^0.23.1",
+    "url-loader": "^2.2.0",
     "webpack": "^4.41.1",
     "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.8.2"
@@ -34,12 +35,14 @@
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.4",
     "express": "^4.17.1",
+    "file-loader": "^4.2.0",
     "node-fetch": "^2.6.0",
     "nodemon": "^1.19.3",
     "pg": "^7.12.1",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
     "react-hot-loader": "^4.12.15",
+    "react-router-dom": "^5.1.2",
     "sequelize": "^5.19.6",
     "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.0"
